+++ packages/react/.eslintrc.js
@@ -4,8 +4,14 @@ module.exports = {
   extends: [
     "@knocklabs/eslint-config/library.js",
     "plugin:react-hooks/recommended",
+    "plugin:jsx-a11y/strict",
   ],
   parserOptions: {
     projects: ["tsconfig.json", "tsconfig.node.json"],
+  },
+  settings: {
+    "jsx-a11y": {
+      polymorphicPropName: "as",
+    },
   },
 };