+++ packages/react/src/modules/feed/components/NotificationIconButton/NotificationIconButton.tsx
@@ -21,12 +21,11 @@ export const NotificationIconButton = React.forwardRef<
   return (
     <button
       className={`rnf-notification-icon-button rnf-notification-icon-button--${colorMode}`}
-      role="button"
       aria-label="Open notification feed"
       ref={ref}
       onClick={onClick}
     >
-      <BellIcon />
+      <BellIcon aria-hidden />
       <UnseenBadge badgeCountType={badgeCountType} />
     </button>
   );